fixup! feature: add support for cookies
